Split app spec creation into several components
ToscaPlanToSpecTransformer loads ApplicationSpecsBuilder class from
Spring platform and uses it to build the specs that are applied to
the root application entity.

ASB uses EntitySpecFactory to pick a spec for each node. The spec is
configured by loading and applying a number of SpecModifier classes.
Each SpecModifier makes a logically distinct change.

Integration tests now run in a Spring test context.