Config resolving: handle DSL values
The config value coercion broke when Brooklyn was changed to respect
generics (e.g. a config value of type DSL cannot be coerced to a String,
before it is evaluated - but this is when we’re creating the EntitySpec
so we can’t evaluate it yet).

Also more testing: a simple config key of type String could previously
not be set with a DSL value.